pyproject.toml
[tool.poetry]
name = "interruption-template"
version = "0.1.0"
description = "A LangGraph application demonstrating dynamic interruption and state management."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.1.1"
langchain = "^0.2.7"
langchain-openai = "^0.1.14"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
langchain-community = "^0.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
